+++ schemacrawler-ai-langchain4j/src/main/java/schemacrawler/tools/command/aichat/langchain4j/Langchain4JToolExecutor.java
@@ -1,4 +1,4 @@
-package schemacrawler.tools.command.aichat.utility.langchain4j;
+package schemacrawler.tools.command.aichat.langchain4j;
 
 import java.sql.Connection;
 import static java.util.Objects.requireNonNull;
