+++ schemacrawler-ai-langchain4j/src/main/java/schemacrawler/tools/command/aichat/langchain4j/Langchain4JUtility.java
@@ -26,7 +26,7 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.tools.command.aichat.utility.langchain4j;
+package schemacrawler.tools.command.aichat.langchain4j;
 
 import java.sql.Connection;
 import java.util.ArrayList;
