+++ schemacrawler-ai-langchain4j/src/main/java/schemacrawler/tools/command/aichat/langchain4j/OpenAIModelFactory.java
@@ -26,33 +26,35 @@ http://www.gnu.org/licenses/
 ========================================================================
  */
 
-package schemacrawler.tools.command.aichat.utility.langchain4j;
+package schemacrawler.tools.command.aichat.langchain4j;
 
 import java.time.Duration;
 import static java.util.Objects.requireNonNull;
-import dev.langchain4j.exception.UnsupportedFeatureException;
 import dev.langchain4j.memory.ChatMemory;
-import dev.langchain4j.memory.chat.MessageWindowChatMemory;
-import dev.langchain4j.model.anthropic.AnthropicChatModel;
-import dev.langchain4j.model.anthropic.AnthropicChatModelName;
+import dev.langchain4j.memory.chat.TokenWindowChatMemory;
 import dev.langchain4j.model.chat.ChatModel;
 import dev.langchain4j.model.embedding.EmbeddingModel;
+import dev.langchain4j.model.openai.OpenAiChatModel;
+import dev.langchain4j.model.openai.OpenAiChatModelName;
+import dev.langchain4j.model.openai.OpenAiEmbeddingModel;
+import dev.langchain4j.model.openai.OpenAiEmbeddingModelName;
+import dev.langchain4j.model.openai.OpenAiTokenCountEstimator;
 import schemacrawler.tools.command.aichat.options.AiChatCommandOptions;
-import schemacrawler.tools.command.aichat.utility.langchain4j.AiModelFactoryUtility.AiModelFactory;
+import schemacrawler.tools.command.aichat.langchain4j.AiModelFactoryUtility.AiModelFactory;
 import us.fatehi.utility.property.PropertyName;
 
-public class AnthropicModelFactory implements AiModelFactory {
+public class OpenAIModelFactory implements AiModelFactory {
 
-  private final PropertyName aiProvider = new PropertyName("anthropic", "Anthropic");
+  private final PropertyName aiProvider = new PropertyName("openai", "OpenAI");
   private final AiChatCommandOptions aiChatCommandOptions;
 
-  public AnthropicModelFactory(final AiChatCommandOptions commandOptions) {
+  public OpenAIModelFactory(final AiChatCommandOptions commandOptions) {
     aiChatCommandOptions = requireNonNull(commandOptions, "No AI Chat options provided");
   }
 
   @Override
   public boolean hasEmbeddingModel() {
-    return false;
+    return true;
   }
 
   @Override
@@ -61,21 +63,23 @@ public class AnthropicModelFactory implements AiModelFactory {
       return false;
     }
     final String model = aiChatCommandOptions.model();
-    for (final AnthropicChatModelName modelName : AnthropicChatModelName.values()) {
+    for (final OpenAiChatModelName modelName : OpenAiChatModelName.values()) {
       if (modelName.toString().equals(model)) {
         return true;
       }
     }
-    return false;
+    return model.startsWith("gpt-");
   }
 
   @Override
   public ChatModel newChatModel() {
-    return AnthropicChatModel.builder()
+    return OpenAiChatModel.builder()
         .apiKey(aiChatCommandOptions.apiKey())
         .modelName(aiChatCommandOptions.model())
         .temperature(0.2)
         .timeout(Duration.ofSeconds(aiChatCommandOptions.timeout()))
+        // https://docs.langchain4j.dev/integrations/language-models/open-ai#structured-outputs-for-tools
+        .strictTools(true)
         .logRequests(true)
         .logResponses(true)
         .build();
@@ -83,12 +87,18 @@ public class AnthropicModelFactory implements AiModelFactory {
 
   @Override
   public ChatMemory newChatMemory() {
-    return MessageWindowChatMemory.withMaxMessages(aiChatCommandOptions.context());
+    return TokenWindowChatMemory.builder()
+        .maxTokens(8_000, new OpenAiTokenCountEstimator(aiChatCommandOptions.model()))
+        .build();
   }
 
   @Override
   public EmbeddingModel newEmbeddingModel() {
-    throw new UnsupportedFeatureException("Anthropic does not have embedding models");
+    final String embeddingModelName = OpenAiEmbeddingModelName.TEXT_EMBEDDING_3_SMALL.toString();
+    return OpenAiEmbeddingModel.builder()
+        .apiKey(aiChatCommandOptions.apiKey())
+        .modelName(embeddingModelName)
+        .build();
   }
 
   @Override
