+++ schemacrawler-ai-langchain4j/src/main/resources/META-INF/services/schemacrawler.tools.command.aichat.ChatAssistant
@@ -1,1 +1,1 @@
-schemacrawler.tools.command.aichat.utility.langchain4j.Langchain4JChatAssistant
+schemacrawler.tools.command.aichat.langchain4j.Langchain4JChatAssistant
